Augmenting :PluginList output with the "name"s
The other Vundle commands that take a plugin name as an argument are
NOT expecting the same string used in the .vimrc file's "Plugin ..."
commands.  I cannot always remember what they expect, and with this
change I don't have to.  I can just do a listing, and the last entry on
each line is exactly what I need to provide to the other Vundle
commands.