Commiting 272d9fe from "incomplete" PR 604
This change ended up on a 'revert-616-fix-github-url', which was never
merged into master.  Now that I'm forking & tweaking the plugins I use,
this change is important to me.

For now, my plan is to keep my changes in a work branched named
'neb-dev', and leave the master alone so I can more easily stay in sync
with the upstream repo's.  Then the only actual merges would then be
from my own 'master' to my work branch.