Add fix for when 'master' is not the default
My last update added the ability to specify the desired remote branch
when setting up the Plugin line in the .vimrc file.

That code I found from an unmerged work branch in the original repo made
the assumption that 'master' would always be the default branch.  It
turns out some of my plugins have 'main' as the default remote branch,
and no 'master' branch, which shows up as errors during a PluginUpdate

This change removes that assumption.